requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
Django==4.2.7
django-anymail==10.2
django-cors-headers==4.3.0
django-environ==0.11.2
django-filter==23.3
django-js-asset==2.1.0
django-map-widgets==0.4.2
django-mptt==0.15.0
django-polymorphic==3.1.0
django-polymorphic-tree==2.1
django-render-block==0.9.2
django-rest-polymorphic==0.1.10
django-tag-parser==3.2
django-templated-email==3.0.1
djangorestframework==3.14.0
djangorestframework-gis==1.0
frozenlist==1.4.0
h3==4.0.0b2
idna==3.4
Markdown==3.5.1
multidict==6.0.4
Pillow==10.1.0
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg-pool==3.1.8
pycparser==2.21
Pygments==2.17.0
PyJWT==2.8.0
python-monkey-business==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
twilio==8.10.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
yarl==1.9.2
